requirements.txt
# Main requirements for Vocabulary Recommendation System
# Data generation and processing
numpy>=1.19.0
requests>=2.32.0
pydantic>=2.0.0

# AWS integration
boto3>=1.34.0
botocore>=1.34.0
PyJWT>=2.0.0

# OpenAI integration for enhanced recommendations
openai>=1.0.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
